refactor: move custom category component
